Moving namespaces around and fixing some stuff
