+++ common/codeception.yml
@@ -11,6 +11,10 @@ settings:
     colors: true
     memory_limit: 1024M
     log: true
+modules:
+  config:
+    Yii2:
+      configFile: 'config/test-local.php'
 coverage:
   enabled: true
   whitelist:
