+++ common/config/test.php
@@ -1,0 +1,11 @@
+<?php
+return [
+  'id' => 'app-common-tests',
+  'basePath' => dirname(__DIR__),
+  'components' => [
+    'user' => [
+      'class' => 'yii\web\User',
+      'identityClass' => 'common\models\User',
+    ],
+  ],
+];
