+++ common/tests/_bootstrap.php
@@ -1,7 +1,10 @@
 <?php
 defined('YII_DEBUG') or define('YII_DEBUG', true);
 defined('YII_ENV') or define('YII_ENV', 'test');
-defined('YII_APP_BASE_PATH') or define('YII_APP_BASE_PATH', __DIR__.'/../');
-require_once(__DIR__ .  '/../../vendor/autoload.php');
-require_once(__DIR__ .  '/../../vendor/yiisoft/yii2/Yii.php');
+defined('YII_APP_BASE_PATH') or define('YII_APP_BASE_PATH', dirname(dirname(__DIR__)));
+
+require_once(YII_APP_BASE_PATH . '/vendor/autoload.php');
+require_once(YII_APP_BASE_PATH . '/vendor/yiisoft/yii2/Yii.php');
+require(YII_APP_BASE_PATH . '/common/config/aliases.php');
+
 Yii::setAlias('@tests', dirname(__DIR__));
