+++ site/config/test.php
@@ -1,0 +1,10 @@
+<?php
+return [
+  'id' => 'app-frontend-tests',
+  'components' => [
+    'user' => [
+      'class' => 'yii\web\User',
+      'identityClass' => 'common\models\User',
+    ],
+  ],
+];
