+++ site/tests/unit.suite.yml
@@ -4,3 +4,8 @@
 # RUN `build` COMMAND AFTER ADDING/REMOVING MODULES.
 
 class_name: UnitTester
+modules:
+  enabled:
+    - Asserts
+    - Yii2:
+      part: ['init', 'orm', 'fixtures', 'email']
